requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-debug-toolbar==5.2.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
python-dotenv